16 octubre

+++ borrar.php
@@ -1,12 +1,14 @@
 <?php 
 
 //$id = trim($_POST['id']); 
+require 'auxiliar.php';
 
 $id = filter_input(INPUT_POST, 'id', FILTER_VALIDATE_INT, FILTER_NULL_ON_FAILURE);
 
-$pdo = new PDO('pgsql:host=localhost;dbname=steam', 'steam', 'steam');
-$sent = $pdo->prepare("DELETE FROM cliente WHERE id = :id");
-$sent->execute([':id' => $id]); 
+if ($id){
+    $pdo = conectar();
+    $sent = $pdo->prepare("DELETE FROM cliente WHERE id = :id");
+    $sent->execute([':id' => $id]); 
+}
 
 header('Location: index.php');
-exit;
